requirements.txt
pipecat-ai[google,deepgram,daily]==0.0.87
google-generativeai>=0.8.5
deepgram-sdk>=3.1.10
chromadb>=0.5.10
websockets>=15.0.1
python-dotenv>=1.1.1
sqlalchemy>=2.0.43
pyyaml>=6.0
sentence-transformers>=3.3.1
transformers>=4.47.0
accelerate>=1.2.1
dspy-ai>=2.5.57
aiohttp>=3.9.0
pipecat-ai-small-webrtc-prebuilt>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
apscheduler>=3.11.0
fastapi-sso>=0.18.0
